Fix torque in observation
According to the documentation, pyBullet only reports actual torques in
`getJointStates` if position or velocity control is used.  In torque
control mode the applied torque is exactly what is provided by the user,
so they don't report it separately but just set it to zero here...

As a fix for this, store the applied torque of the current step in a
variable and use this to fill the observation.